refactor: replace CustomDedent with Dedent from Formatters for improved string handling
