+++ Source/Utilities/Strings/Formatters.ts
@@ -7,6 +7,7 @@ import { userMention } from "discord.js";
 import { TitleCase } from "./Converters.js";
 import { Vehicles } from "@Typings/Resources.js";
 import ERLCAgeGroups from "@Resources/ERLCAgeGroups.js";
+import _Dedent from "dedent";
 
 /**
  * Formats an input multiline string of charges into a properly formatted numbered list.
@@ -746,6 +747,23 @@ export function FormatSortRDInputNames(
     else if (IsValidRobloxUsername(Name)) return AddRUsernamePrefix ? `@${Name}` : Name;
     else return Name;
   });
+}
+
+/**
+ * Removes excess unwanted indentation and extra spaces from a given string potentially caused by escaping newlines in a multiline string.
+ *
+ * This function processes the input string to dedent it and then performs
+ * additional cleanup by replacing:
+ * - Periods followed by two or more spaces with a single space after the period.
+ * - Words separated by two or more spaces with a single space.
+ *
+ * @param text - The input string to be dedented and cleaned.
+ * @returns The processed string with reduced indentation and extra spaces removed.
+ */
+export function Dedent(text: string): string {
+  return _Dedent(text)
+    .replace(/\.[^\S\r\n]{2,}(\w)/g, ". $1")
+    .replace(/(\w)[^\S\r\n]{2,}(\w)/g, "$1 $2");
 }
 
 /**
